add sweet alert lib and change webpack and move styles to folder(css_modules)
