+++ src/css_modules/app.css
@@ -10,4 +10,5 @@
   right: 0;
   bottom: 0;
   top: 0;
+  overflow-y: scroll;
 }
