+++ src/css_modules/statistic.css
@@ -4,7 +4,6 @@
   grid-template-columns: auto;
   align-items: center;
   justify-items: center;
-  margin-bottom: 400px;
 }
 
 @media screen and (max-width: 700px) {
